Created a liquidity bot for bitfloor.
This bot will check the order book, and place an order thats 30 cents
above/below the 2nd highest/lowest bid/ask respectively.  It checks to
make sure that its the only resident of that price point. This way you
can profit off the 0.1% provider bonus on orders placed on the book on
bitfloor. I'm starting a 2nd method with new logic of this as well.