Added a "buy/sell usd <#USD> [price]" option
Now you can specify your buy and sell amounts in usd. both "buy usd 50"
and "buy 50 usd" works.... if no price is given, it will go as a market
order at current rates (ticker "buy" or "sell" at the time of command).
This may overshoot or undershoot the target if the market moves (lag) or
there is not enough available BTC to fulfill your order at the best
price.